+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+* Better unicode support in `Konva.Text` and `Konva.TextPath`. Emoji should work better now 👍
+
 ## 7.1.0
 
 * Multi row support for `ellipsis` config for `Konva.Text`